requirements.txt
blessed==1.20.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
fsspec==2023.12.2
googletrans==4.0.0rc1
gTTS==2.4.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inquirer==3.1.4
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai-whisper==20231117
python-editor==1.0.4
readchar==4.0.5
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sympy==1.12
tiktoken==0.5.2
torch==2.1.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
